Restriction to pan
Added the option to restrict panning (dragging behaviour) to the initial date range. So if the date range is from 2010 to 2012, it is not possible to pan to say 2009.
If the minimumScale is less than 1, it is possible to zoom out, however panning will be disabled so that it is not possibel to pan to say an earlier date.
ZoomToDomain function will ignore this (this is default D3 behaviour).